refactored UI code
